refactor: 삭제상태 추가해서 소프트 딜리트로 수정

+++ src/main/java/com/gongspot/project/domain/newplace/dto/NewPlaceResponseDTO.java
@@ -29,8 +29,8 @@ public class NewPlaceResponseDTO {
     @AllArgsConstructor
     @Builder
     public static class NewProposalHomeDTO {
-        private Integer totalAllProposalsCount;
-        private Integer totalUnapprovedProposalsCount;
+        private long totalAllProposalsCount;
+        private long totalUnapprovedProposalsCount;
         private List<NewProposalDTO> unapprovedProposals;
     }
 }
